requirements.txt
albumentations==1.4.2
comet_ml==3.39.2
coremltools==7.1
easydict==1.13
Flask==3.0.2
GitPython==3.1.42
gradio==4.24.0
ipython==8.12.3
matplotlib==3.7.0
mss==9.0.1
nltk==3.8.1
numpy==1.23.5
onnx==1.13.1
onnxruntime==1.14.0
onnxsim==0.4.36
opencv_contrib_python==4.7.0.72
opencv_python==4.7.0.72
openvino==2024.0.0
paddle==1.0.2
pafy==0.5.5
pandas==1.5.3
Pillow==9.4.0
prettytable==3.10.0
psutil==5.9.8
PyYAML==6.0.1
Requests==2.31.0
ruamel.base==1.0.0
scikit_learn==1.4.0
scipy==1.12.0
seaborn==0.13.2
setuptools==67.3.2
tensorflow==2.16.1
tensorflowjs==4.17.0
tensorrt==8.6.1.post1
tflite_runtime==2.14.0
tflite_support==0.4.4
thop==0.1.1.post2209072238
timm==0.6.7
torch==2.1.2
torchvision==0.14.1
tqdm==4.64.1
train==0.0.5
transformers==4.32.0
tritonclient==2.44.0
x2paddle==1.4.1
